Unify websocket server session's message handling
