+++ src/touchpad.js
@@ -9,16 +9,6 @@ var MAX_SPEED = 5;
 // - remote mouse control
 // - server mouse reporting to transmitter
 
-function abToType(b, Type) {
-	var ab = b.buffer.slice(b.byteOffset, b.byteOffset + b.byteLength);
-	return ab;
-}
-
-function fromArrayBuffer(ab, Type) {
-	var array = new Type(ab);
-	return array;
-}
-
 class TouchPadSession {
 	constructor(ws, sessions, roles) {
 		this.ws = ws;
@@ -26,9 +16,6 @@ class TouchPadSession {
 		this.roles = new Set(roles);
 
 		this.speed = MIN_SPEED;
-
-		// ws.on('message', data => this.onMessage(data));
-		// ws.on('close', e => this.onClose(e));
 
 		this.updateMouse();
 
@@ -123,50 +110,6 @@ class TouchPadSession {
 		}
 
 		return 0;
-	}
-
-	onMessage(data) {
-		if (data instanceof Buffer || data instanceof ArrayBuffer) {
-			if (data instanceof Buffer) {
-				var ab = abToType(data);
-			} else {
-				// uws buffer handling
-				var ab = data;
-			}
-			
-			var floats = fromArrayBuffer(ab, Float64Array);
-
-			var cmdCode = floats[0];
-			var cmd = wire.CODES[cmdCode];
-
-			if (cmd === undefined) {
-				console.log('invalid cmd code', cmd);
-			}
-
-			var ts = floats[1];
-			var coords = floats.subarray(2);
-
-			// console.log('meows', cmd, ts, coords);
-			if (this._lag !== undefined) {
-				if (this._lag % 10000 === 0) console.log('lag', Date.now() - ts);
-				this._lag++;
-			} else {
-				this._lag = 0;
-			}
-
-			this.processMessage(cmd, coords, data);
-			return;
-		}
-		else if (typeof data !== 'string') {
-			console.log('Oops unknown data type', typeof data, data);
-			return;
-		}
-
-		var msg = data.split('\n');
-		var cmd = msg[0];
-		var coords = msg[1] && JSON.parse(msg[1]);
-
-		this.processMessage(cmd, coords, data);
 	}
 
 	processMessage(cmd, coords, data) {
